requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
python-dotenv==1.0.1